Split MaybeTimer into a submodule
